code review comments changes
