+++ sdk/python/setup.py
@@ -29,7 +29,6 @@ REQUIRES = [
     "urllib3>=1.15.1",
     "kubernetes>=23.6.0",
     "retrying>=1.3.3",
-    "parameterized>=0.9.0",
 ]
 
 setuptools.setup(
